endpointsharding: shuffle endpoint order before updating children
